Separated Google auth to another tab, refactor admin module
