+++ eform-client/src/app/modules/settings/components/admin-settings/admin-settings.component.html
@@ -3,12 +3,12 @@
     <h1 class="page-title header-page-title">
       <div class="container">
         Settings
-        <h5 style="display: inline" *ngIf="currentRole === 'admin'">
+        <h5 style="display: inline">
           <small>({{adminSettingsModel.assemblyVersion}})</small>
         </h5>
       </div>
     </h1>
-    <div class="pull-right" *ngIf="currentRole === 'admin'">
+    <div class="pull-right">
 
       <button class="btn btn-ar btn-danger" [disabled]="!settingsForm.form.valid || spinnerStatus"
               (click)="updateAdminSettings()">
@@ -20,7 +20,7 @@
 <div class="edit_case">
   <div class="container-fluid">
     <form #settingsForm="ngForm">
-      <ng-container *ngIf="currentRole === 'admin'">
+
         <div class="row" style="padding-bottom: 15px">
           <div class="panel-group">
             <div class="col-md-6" role="main">
@@ -210,11 +210,8 @@
               </div>
             </div>
           </div>
-
         </div>
-      </ng-container>
       <div class="row">
-        <ng-container *ngIf="currentRole === 'admin'">
           <div class="col-md-6" role="main">
             <div class="panel panel-default">
               <div class="panel-heading">
@@ -265,50 +262,6 @@
               </table>
             </div>
           </div>
-        </ng-container>
-
-        <div class="col-md-6" role="main">
-          <div class="panel panel-default">
-            <div class="panel-heading">
-              Google Authenticator
-            </div>
-            <table class="table">
-              <tbody>
-              <tr>
-                <td class="col-md-12">
-                  <div class="checkbox checkbox-wrapper">
-                    <input  name="isTwoFactorEnabled"
-                            id="isTwoFactorEnabled"
-                            value="{{googleAuthInfoModel.isTwoFactorEnabled}}"
-                            [checked]="googleAuthInfoModel.isTwoFactorEnabled"
-                            [disabled]="googleAuthInfoModel.isTwoFactorForced"
-                            (change)="isTwoFactorEnabledCheckBoxChanged($event)"
-                            type="checkbox">
-                    <label for="isTwoFactorEnabled"><strong>Status</strong> (change will logout you from the application)</label>
-                    <br>
-                  </div>
-                  <ng-container *ngIf="googleAuthInfoModel.psk">
-                    <p>Secret key between <strong>Eform</strong> application and <strong>Google Authenticator</strong>.
-                      Keep it safe and do not tell anyone.
-                    </p>
-                    <p>You should remove this key, if <strong>Google Authenticator</strong>
-                      is reinstalled or key was removed</p>
-                    <p><strong>{{googleAuthInfoModel.psk}}</strong></p>
-                  </ng-container>
-                </td>
-              </tr>
-              <tr *ngIf="googleAuthInfoModel.psk">
-                <td class="col-md-12">
-                  <button class="btn btn-ar btn-warning" id="deleteGoogleAuthenticatorInfo"
-                          tooltip="Remove google authenticator secret key" (click)="deleteGoogleAuthenticatorInfo()">
-                    Remove PSK
-                  </button>
-                </td>
-              </tr>
-              </tbody>
-            </table>
-          </div>
-        </div>
       </div>
     </form>
   </div>
